# Add model compilation in ORT perf test