--- a/onnxruntime/test/perftest/command_args_parser.cc
+++ b/onnxruntime/test/perftest/command_args_parser.cc
@@ -169,6 +169,9 @@ ABSL_FLAG(std::string, plugin_ep_options, "",
           "--plugin_ep_options \"ep_1_option_1_key|ep_1_option_1_value ...;;ep_3_option_1_key|ep_3_option_1_value ...;... \"");
 ABSL_FLAG(bool, list_ep_devices, false, "Prints all available device indices and their properties (including metadata). This option makes the program exit early without performing inference.\n");
 ABSL_FLAG(std::string, select_ep_devices, "", "Specifies a semicolon-separated list of device indices to add to the session and run with.");
+ABSL_FLAG(bool, compile_ep_context, DefaultPerformanceTestConfig().run_config.compile_ep_context, "Generate an EP context model");
+ABSL_FLAG(std::string, compile_model_path, "", "The compiled model path for saving EP context model. Overwrites if already exists. Default is model_ctx.onnx");
+ABSL_FLAG(int, compile_embed_mode, DefaultPerformanceTestConfig().run_config.compile_embed_mode, "Embed binary blob within EP context node. Can be 0 or 1. Default is 0");
 ABSL_FLAG(bool, h, false, "Print program usage.");
 
 namespace onnxruntime {
@@ -484,6 +487,18 @@ bool CommandLineParser::ParseArguments(PerformanceTestConfig& test_config, int a
     const auto& select_ep_devices = absl::GetFlag(FLAGS_select_ep_devices);
     if (!select_ep_devices.empty()) test_config.selected_ep_device_indices = select_ep_devices;
   }
+
+  // --compile_ep_context
+  test_config.run_config.compile_ep_context = absl::GetFlag(FLAGS_compile_ep_context);
+
+  // --compile_model_path
+  const auto& compile_model_path = absl::GetFlag(FLAGS_compile_model_path);
+  if (!compile_model_path.empty()) test_config.run_config.compile_model_path = ToPathString(compile_model_path);
+  else test_config.run_config.compile_model_path = ToPathString("model_ctx.onnx");
+
+  // --compile_embed_mode
+  if (absl::GetFlag(FLAGS_compile_embed_mode) < 0 || absl::GetFlag(FLAGS_compile_embed_mode) >= 2) return false;
+  test_config.run_config.compile_embed_mode = absl::GetFlag(FLAGS_compile_embed_mode);
 
   if (positional.size() == 2) {
     test_config.model_info.model_file_path = ToPathString(positional[1]);