--- a/onnxruntime/test/perftest/command_args_parser.cc
+++ b/onnxruntime/test/perftest/command_args_parser.cc
@@ -171,7 +171,7 @@ ABSL_FLAG(bool, list_ep_devices, false, "Prints all available device indices and
 ABSL_FLAG(std::string, select_ep_devices, "", "Specifies a semicolon-separated list of device indices to add to the session and run with.");
 ABSL_FLAG(bool, compile_ep_context, DefaultPerformanceTestConfig().run_config.compile_ep_context, "Generate an EP context model");
 ABSL_FLAG(std::string, compile_model_path, "model_ctx.onnx", "The compiled model path for saving EP context model. Overwrites if already exists");
-ABSL_FLAG(int, compile_embed_mode, DefaultPerformanceTestConfig().run_config.compile_embed_mode, "Embed binary blob within EP context node. Can be 0 or 1");
+ABSL_FLAG(bool, compile_binary_embed, DefaultPerformanceTestConfig().run_config.compile_binary_embed, "Embed binary blob within EP context node");
 ABSL_FLAG(bool, h, false, "Print program usage.");
 
 namespace onnxruntime {
@@ -495,9 +495,8 @@ bool CommandLineParser::ParseArguments(PerformanceTestConfig& test_config, int a
   const auto& compile_model_path = absl::GetFlag(FLAGS_compile_model_path);
   test_config.run_config.compile_model_path = ToPathString(compile_model_path);
 
-  // --compile_embed_mode
-  if (absl::GetFlag(FLAGS_compile_embed_mode) < 0 || absl::GetFlag(FLAGS_compile_embed_mode) >= 2) return false;
-  test_config.run_config.compile_embed_mode = absl::GetFlag(FLAGS_compile_embed_mode);
+  // --compile_binary_embed
+  test_config.run_config.compile_binary_embed = absl::GetFlag(FLAGS_compile_binary_embed);
 
   if (positional.size() == 2) {
     test_config.model_info.model_file_path = ToPathString(positional[1]);