--- a/onnxruntime/test/perftest/main.cc
+++ b/onnxruntime/test/perftest/main.cc
@@ -13,6 +13,11 @@
 using namespace onnxruntime;
 const OrtApi* g_ort = NULL;
 
+
+int RunPerfTest(Ort::Env& env, const perftest::PerformanceTestConfig& test_config);
+void CompileEpContextModel(Ort::Env& env, const perftest::PerformanceTestConfig& test_config);
+
+
 #ifdef _WIN32
 int real_main(int argc, wchar_t* argv[]) {
 #else
@@ -30,7 +35,7 @@ int real_main(int argc, char* argv[]) {
     ORT_TRY {
       OrtLoggingLevel logging_level = test_config.run_config.f_verbose
                                           ? ORT_LOGGING_LEVEL_VERBOSE
-                                          : ORT_LOGGING_LEVEL_WARNING;
+                                          : ORT_LOGGING_LEVEL_ERROR;
       env = Ort::Env(logging_level, "Default");
     }
     ORT_CATCH(const Ort::Exception& e) {
@@ -67,24 +72,24 @@ int real_main(int argc, char* argv[]) {
     return 0;
   }
 
-  std::random_device rd;
-  perftest::PerformanceRunner perf_runner(env, test_config, rd);
+  int status = 0;
 
-  // Exit if user enabled -n option so that user can measure session creation time
-  if (test_config.run_config.exit_after_session_creation) {
-    perf_runner.LogSessionCreationTime();
-    return 0;
+  // EP context perf test
+  if(test_config.run_config.compile_ep_context) {
+    {
+      std::cout << "\n> Compiling model...\n";
+      CompileEpContextModel(env, test_config);
+    }
+
+    {
+      test_config.model_info.model_file_path = test_config.run_config.compile_model_path;
+      status = RunPerfTest(env, test_config);
+    }
+  } else {
+    // regular perf test
+    status = RunPerfTest(env, test_config);
   }
-
-  auto status = perf_runner.Run();
-  if (!status.IsOK()) {
-    printf("Run failed:%s\n", status.ErrorMessage().c_str());
-    return -1;
-  }
-
-  perf_runner.SerializeResult();
-
-  return 0;
+  return status;
 }
 
 #ifdef _WIN32
@@ -106,4 +111,58 @@ int main(int argc, char* argv[]) {
   ::google::protobuf::ShutdownProtobufLibrary();
 
   return retval;
+}
+
+
+int RunPerfTest(Ort::Env& env, const perftest::PerformanceTestConfig& test_config) {
+  std::random_device rd;
+  perftest::PerformanceRunner perf_runner(env, test_config, rd);
+
+  // Exit if user enabled -n option so that user can measure session creation time
+  if (test_config.run_config.exit_after_session_creation) {
+    perf_runner.LogSessionCreationTime();
+    return 0;
+  }
+
+  auto status = perf_runner.Run();
+  if (!status.IsOK()) {
+    printf("Run failed:%s\n", status.ErrorMessage().c_str());
+    return -1;
+  }
+
+  perf_runner.SerializeResult();
+  return 0;
+}
+
+
+void CompileEpContextModel(Ort::Env& env, const perftest::PerformanceTestConfig& test_config) {
+  auto output_ctx_model_path = test_config.run_config.compile_model_path;
+  const auto provider_name_ = test_config.machine_config.provider_type_name;
+
+  Ort::SessionOptions session_options;
+
+  std::unordered_map<std::string, std::string> provider_options_;
+  session_options.AppendExecutionProvider(provider_name_, provider_options_);
+
+  Ort::ModelCompilationOptions model_compile_options(env, session_options);
+  model_compile_options.SetEpContextEmbedMode(test_config.run_config.compile_embed_mode);
+  model_compile_options.SetInputModelPath(test_config.model_info.model_file_path.c_str());
+  model_compile_options.SetOutputModelPath(output_ctx_model_path.c_str());
+
+  Ort::Status status;
+  std::chrono::duration<double> compile_duration;
+  {
+    auto compile_time_start_ = std::chrono::high_resolution_clock::now();
+    status = Ort::CompileModel(env, model_compile_options);
+    auto compile_time_end_ = std::chrono::high_resolution_clock::now();
+    compile_duration = compile_time_end_ - compile_time_start_;
+  }
+  model_compile_options.release();
+
+  if (!status.IsOK()) {
+    std::cout << "Failed to compile model: " << status.GetErrorMessage() << std::endl;
+  } else {
+    std::cout << "Compile time cost: " << compile_duration.count() << " s\n";
+  }
+  return;
 }