@edgchen1 2025-09-15
> > why is this PR adding two ways to compile the model?
> 
> The core utility is the same. Just added the same implementation for Python and C++ to reduce dependency overhead and devs can use whichever language they are already working on. Python is quicker to test and prototype.

I haven't used this, but a potential concern is having two separate implementations to maintain. is having both paths worth that maintenance overhead?